+++ charts/external-dns/CHANGELOG.md
@@ -22,6 +22,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Avoid creating cluster-scoped RBAC for Gateway API sources when running namespaced with `gatewayNamespace` set. Namespace listing permissions are now only added when `gatewayNamespace` is unset. ([#5843](https://github.com/kubernetes-sigs/external-dns/pull/5843)) _@TobyTheHutt_
 
+### Changed
+
+- Use image tag and digest by default to ensure immutability of the image used by the chart.
+  ([#6355](https://github.com/kubernetes-sigs/external-dns/pull/6355)) _@vflaux_
+
 ## [v1.20.0]
 
 ### Added